server-requirements.txt
alembic==1.13.2
babel==2.16.0
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
boto3==1.35.10
botocore==1.35.10
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
Flask==3.0.3
flask-babel==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gevent==25.4.2
gevent-websocket==0.10.1
greenlet==3.2.1
gunicorn==23.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
qrcode==7.4.2
redis==5.0.8
rq==1.16.2
s3transfer==0.10.2
setuptools==80.1.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
zope.event==5.0
zope.interface==7.2

